# docs/swagger.yaml
basePath: /
definitions:
  model.BaseResponse:
    properties:
      data: {}
      description:
        type: string
      message:
        type: string
      pagination:
        $ref: '#/definitions/model.Pagination'
      status:
        type: string
      success:
        type: boolean
    type: object
  model.CourseCreateRequest:
    properties:
      description:
        type: string
      name:
        type: string
      photo:
        type: string
    required:
    - name
    type: object
  model.CourseUpdateRequest:
    properties:
      description:
        type: string
      name:
        type: string
      photo:
        type: string
    required:
    - name
    type: object
  model.ExamTypeCreateRequest:
    properties:
      max_ball:
        type: number
      name:
        type: string
    required:
    - name
    type: object
  model.ExamTypeUpdateRequest:
    properties:
      max_ball:
        type: number
      name:
        type: string
    type: object
...
  model.LoginRequest:
    properties:
      password:
        default: P@$$w0rd2o24
        type: string
      username:
        default: superadmin
        type: string
    type: object
  model.Pagination:
    properties:
      item_total:
        type: integer
      page:
        default: 1
        type: integer
      page_size:
        default: 10
        type: integer
      page_total:
        type: integer
    type: object
  model.RefreshRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Server for Application
  title: EduSync API
  version: "1.0"
paths:
  /api/courses:
    get:
      consumes:
      - application/json
      description: Get List Course
      parameters:
      - default: 10
        description: pageSize
        in: query
        name: pageSize
        required: true
        type: integer
      - default: 1
        description: page
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Get List Course
      tags:
      - Course
    post:
      consumes:
      - application/json
      description: Create Course
      parameters:
      - description: Create Course
        in: body
        name: create
        required: true
        schema:
          $ref: '#/definitions/model.CourseCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Course
      tags:
      - Course
  /api/courses/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Course
      parameters:
      - description: Course Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Course
      tags:
      - Course
    get:
      consumes:
      - application/json
      description: Get Course By Id
      parameters:
      - description: Course Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Course By Id
      tags:
      - Course
    put:
      consumes:
      - application/json
      description: Update Course
      parameters:
      - description: Course Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update Course
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/model.CourseUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Course
      tags:
      - Course
  /api/courses/{id}/exam-types:
    get:
      consumes:
      - application/json
      description: Get List Course Exam Type
      parameters:
      - description: Course Id
        in: path
        name: id
        required: true
        type: integer
      - default: 10
        description: pageSize
        in: query
        name: pageSize
        required: true
        type: integer
      - default: 1
        description: page
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Get List Course Exam Type
      tags:
      - Course
    post:
      consumes:
      - application/json
      description: Create Course Exam Type
      parameters:
      - description: Course Id
        in: path
        name: id
        required: true
        type: integer
      - description: Create Course
        in: body
        name: create
        required: true
        schema:
          $ref: '#/definitions/model.ExamTypeCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Course Exam Type
      tags:
      - Course
  /api/courses/{id}/exam-types/{examTypeId}:
    delete:
      consumes:
      - application/json
      description: Delete Course Exam Type
      parameters:
      - description: Course Id
        in: path
        name: id
        required: true
        type: integer
      - description: Exam Type Id
        in: path
        name: examTypeId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Course Exam Type
      tags:
      - Course
    get:
      consumes:
      - application/json
      description: Get List Course Exam Type
      parameters:
      - description: Course Id
        in: path
        name: id
        required: true
        type: integer
      - description: Exam Type Id
        in: path
        name: examTypeId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Get List Course Exam Type
      tags:
      - Course
    put:
      consumes:
      - application/json
      description: Update Course Exam Type
      parameters:
      - description: Course Id
        in: path
        name: id
        required: true
        type: integer
      - description: Exam Type Id
        in: path
        name: examTypeId
        required: true
        type: integer
      - description: Update Course
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/model.ExamTypeUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Course Exam Type
      tags:
      - Course
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login User
      parameters:
      - description: Login
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      summary: Login User
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh Token
      parameters:
      - description: Refresh Token
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/model.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BaseResponse'
      summary: Refresh Token
      tags:
      - Auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"